Kneedle Update and Gigaloop Formatting
Sample Code for a Kneedle-algorithm inspired method of detecting knees in our hyperloop  data was added. The main version takes a hyperloop output and calculates the knee. If this spot is to be analysed in more detail, a secondary bit of Kneedle code was added, which allows you to analyse more high resolution hyperloop data from that spot using previous kneedle.line statistics. Finally, the Gigaloop Code was moved from "Data Loading" into this file, as it is not code for Data Loading, but for drawing a plot. Other edits are simple formatting for increased legibility.