Streamlining of the code
Creating the dataset g_patients which includes just coverage and beta values of the genes. Deleting of datasets with only one group of patients and either beta values or coverage values. Deletion of joined datasets with different patients with average coverage value for each gene.  Shortening of the code lines for removing lines in g_patients with more than for NAs. Optimizing the loops for converting beta values into M values for promoters and genes by removing unnecessary backup datasets and renaming of g/p_normalization into g/p_Mvalues. Removing of unnecessary datasets which are not required for the downstream process at the end of the code.