Create Code for  Imputation Models.R
This code was developed to study the relationship between different imputation methods and the false positive rate.

Two seven value samples (that are identically distributed) are taken and each sample hast 3 values imputed into it in one of five ways. These samples are compared with a t test. Once this process has been repeated 2 million times for each type of imputation, a histogram of p values is drawn up for every method. More details in the code.