Create Code for Data Structure in rnorm10 and Mean Imputation.R
This code analyses two imputation methods (rnorm10and mean) and compares them too each other to analyse which of the two emulates the internal data structure of the original data the best.

This is done by first creating original data and then losing some of it's values. These NAs are then imputed perfectly, via mean, and via rnorm10.  (In the perfect scenario the NAs are simply imputed by the values that were cut).

The actual test consists of a t-test comparing the imputed vector to the original vector, a wilcoxon test doing the same thing, and the average difference between imputed and real values . In this sense mean, sd, rank of ordered values and  distance between imputed and real values is calculated and quantified.

Once this model runs through 500.000 iterations the average results are projected to the use.