Moved Code for PCA Analysis into new file
This code authored by Konstantin Fisher is integral for the PCA analysis of our data as well as visualizing data structures in our PCs with plots and creating dataframes for the further downstream analysis of batch effects.

As it is an anlysis in its own right it was taken out of the data loading and transformation file.