Create Code for DMR search.R
This code relies on names of the genes from PC1 that we want to include, these should be named g_PC1_names and p_PC1__names (provided that for prommoters PC1 is taken later on).

This code requires the qvalue package (https://github.com/StoreyLab/qvalue) to function.

It first takes the T data sets and uses their t-test input valuesd to calculate p values for every sequence manually, without relying on imputation.

In addition, once these p values are calculated a histogramm of them is produced to acertain wether or not they are usable with the q value method.

Q values are calculated for all post QC sequences, to give the most realistic and conventional application of q values.

Frrom these q value assosciated data sets, sequences considered relevant via PCA analysis are extracted.

These data sets are merged with the resource data set to assosciate every sequence with fc, symbol and searchable ensign id.

Of these, those with q <= 0.05 are removed (~ p = 0.035) and those with a fc in a range of 3/4 to 4/3  are removed (low biological significance).

These final candidates are ranked by a mixture of significance and fold change, and are then plotted in comparison to the full post qc data set, with lines marking which areas are boundaries for statistical and biological significance.