Replaced Code for Clustering Analysis (Update)
Requires g_Mvalues and g_pca to run.

This is a rewrite or edit on the original clustering code that is easier to edit and use do to condensing repetitive code into one apply loop.

Also adds a function that describes the accuracy of kmeans clustering for varrying degrees of PC1's loadings.

Also added progress bars and removed some uneccesary formatting steps.